ewmh: icon data are longs on 64-bit hosts
On 64-bit hosts XLib returns 32-bit values in a long type, not an int.  Notably
this causes crashes when handling image data which is misinterpreted.

Fix this by changing the relevant types to long, and using the new
_from_long_data() function in libmatchbox.

+++ src/ewmh.h
@@ -87,8 +87,8 @@ void
 ewmh_hung_app_check (Wm *w);
 
 #ifndef REDUCE_BLOAT
-int 
-*ewmh_get_icon_prop_data (Wm *w, Window win);
+unsigned long *
+ewmh_get_icon_prop_data (Wm *w, Window win);
 #endif
 
 int
